% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUE2.R
\name{DUE2}
\alias{DUE2}
\title{DUE2 function}
\usage{
DUE2(
  graph,
  data,
  group,
  glm.family,
  alpha = 0.05,
  method.FDR = "BH",
  method.FWER = "holm",
  cell.group = NULL,
  sf = NULL,
  progressbar = TRUE
)
}
\arguments{
\item{graph}{A symmetric adjacency matrix representing the undirected graph. Node names must match the variable names in the data.}

\item{data}{A data frame with the observed variables. Variable names in the data must match the node names.}

\item{group}{A vector of group labels for comparison, one entry per observation.}

\item{glm.family}{Description of the distribution and link function to be used in the model.}

\item{alpha}{The alpha level for conducting the test. Default is 0.05.}

\item{method.FDR}{Method for adjusting p-values in the first stage to control the FDR. Default is 'BH'.}

\item{method.FWER}{Method for adjusting p-values in the first stage to control the FWER. Default is 'holm'.}

\item{cell.group}{A vector of cell group labels, used as an extra covariate in the model and not for comparison, with length equal to the sample size.}

\item{sf}{A vector of size factors with length equal to the sample size. It is used as an offset and to scale the covariates in the model.}

\item{progressbar}{Logical, whether to display the progress bar. Default is TRUE.}
}
\value{
a list with four elements.
1. diff: A matrix of dimension p x (p + 2), where p is the number of nodes, indicating differences in the network.
The columns are as follows:
1st column: First screening stage: 0 for equality, 1 for nodes with a difference.
2nd column: Difference in the mean parameter of the node in the row: 0 for no difference, 1 for a difference, NA if the first column is 0.
3-(p+2)th columns: Differences in the association parameters between the node in the row and nodes in the columns: 0 for no difference, 1 for a difference, NA if not tested.

2. delta_pvalues: corrected p-values of the two stages.

3. delta_coef: coefficients of the difference parameters.

4. base_coef: coefficients of the reference condition.
}
\description{
Function to identify differences in networks between two conditions.
}
\details{
1. Screening stage

2. Confirmatory stage
}
