% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_nb.R
\name{test_nb}
\alias{test_nb}
\title{Local regressions and tests - Negative Binomial}
\usage{
test_nb(graph, data, group, glm.family, cell.group, sf, progressbar)
}
\arguments{
\item{graph}{A symmetric adjacency matrix representing the undirected graph. Node names must match the variable names in the data.}

\item{data}{A data frame with the observed variables. Variable names in the data must match the node names.}

\item{group}{A vector of group labels for comparison, one entry per observation.}

\item{glm.family}{Description of the distribution and link function to be used in the model.}

\item{cell.group}{A vector of cell group labels, used as an extra covariate in the model and not for comparison, with length equal to the sample size.}

\item{sf}{A vector of size factors with length equal to the sample size. It is used as an offset and to scale the covariates in the model.}

\item{progressbar}{Logical, whether to display the progress bar. Default is TRUE.}
}
\value{
A list containing three elements: a matrix of p-values, a matrix of coefficients for the baseline group, and a matrix of coefficients for the differences between the two conditions.
}
\description{
Function to fit node-conditional models for comparison and calculate p-values for the two stages.
}
